# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example Project",
        "version": "3.0.1"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User Login",
                "description": "Token almak için kullanıcı girişi",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "email@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/items": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "List All Items",
                "operationId": "2cbf07eb95d033f8775c36424ac8fa06",
                "responses": {
                    "200": {
                        "description": "İtemleri listeler.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/item": {
            "put": {
                "tags": [
                    "Items"
                ],
                "summary": "Update Item",
                "description": "Item düzenle",
                "operationId": "ae1a177ee6f7b2897d822fbfc78c32ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "name",
                                    "description",
                                    "quantity"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "123"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "description"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Create Item",
                "description": "Item oluştur",
                "operationId": "5567f92eba0a6eaacf0124498e81dadf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "quantity"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "description"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/item/{id}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Show Item",
                "description": "Item görüntüle",
                "operationId": "c6650aff4f52c233e3fd6e521dfc2be4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Items"
                ],
                "summary": "Delete Items",
                "description": "Item siler",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "apiKeyAuth": []
        }
    ],
    "tags": [
        {
            "name": "User",
            "description": "User"
        },
        {
            "name": "Items",
            "description": "Items"
        }
    ]
}